try fix cast from 'bool *' to 'int *'

+++ tasks/all/kholin_k_multidimensional_integrals_rectangle/func_tests/func_all.cpp
@@ -15,7 +15,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, test_validation) {
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 1;
+  int dim = 1;
   std::vector<double> values{0.0};
   auto f = [](const std::vector<double> &f_values) { return std::sin(f_values[0]); };
   std::vector<double> in_lower_limits{0};
@@ -47,7 +47,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, test_pre_processing) {
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 1;
+  int dim = 1;
   std::vector<double> values{0.0};
   auto f = [](const std::vector<double> &f_values) { return std::sin(f_values[0]); };
   std::vector<double> in_lower_limits{0};
@@ -80,7 +80,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, test_run) {
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 1;
+  int dim = 1;
   std::vector<double> values{0.0};
   auto f = [](const std::vector<double> &f_values) { return std::sin(f_values[0]); };
   std::vector<double> in_lower_limits{0};
@@ -114,7 +114,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, test_post_processing) {
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 1;
+  int dim = 1;
   std::vector<double> values{0.0};
   auto f = [](const std::vector<double> &f_values) { return std::sin(f_values[0]); };
   std::vector<double> in_lower_limits{0};
@@ -149,7 +149,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, single_integral_one_var)
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 1;
+  int dim = 1;
   std::vector<double> values{0.0};
   auto f = [](const std::vector<double> &f_values) { return f_values[0]; };
   std::vector<double> in_lower_limits{2};
@@ -189,7 +189,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, single_integral_two_var)
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 1;
+  int dim = 1;
   std::vector<double> values{0.0, 3.0};
   auto f = [](const std::vector<double> &f_values) { return f_values[0] + f_values[1]; };
   std::vector<double> in_lower_limits{0};
@@ -230,7 +230,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, double_integral_two_var)
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 2;
+  int dim = 2;
   std::vector<double> values{0.0, 0.0};
   auto f = [](const std::vector<double> &f_values) { return (2 * f_values[0]) + (2 * f_values[1]); };
   std::vector<double> in_lower_limits{0, 0};
@@ -271,7 +271,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, double_integral_one_var)
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 2;
+  int dim = 2;
   std::vector<double> values{-17.0, 0.0};
   auto f = [](const std::vector<double> &f_values) { return 289 + (f_values[1] * f_values[1]); };
   std::vector<double> in_lower_limits{-10, 3};
@@ -312,7 +312,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, triple_integral_three_va
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 3;
+  int dim = 3;
   std::vector<double> values{0.0, 0.0, 0.0};
   auto f = [](const std::vector<double> &f_values) {
     return (f_values[0] * f_values[0]) + (f_values[1] * f_values[1]) + (f_values[2] * f_values[2]);
@@ -355,7 +355,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, triple_integral_two_var)
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 3;
+  int dim = 3;
   std::vector<double> values{0.0, 5.0, 0.0};
   auto f = [](const std::vector<double> &f_values) { return (f_values[0] + f_values[1]); };
   std::vector<double> in_lower_limits{0, 0, 0};
@@ -396,7 +396,7 @@ TEST(kholin_k_multidimensional_integrals_rectangle_all, triple_integral_one_var)
   int rank = 0;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   // Create data
-  size_t dim = 3;
+  int dim = 3;
   std::vector<double> values{0.0, 5.0, -10.0};
   auto f = [](const std::vector<double> &f_values) { return f_values[0] + 5.0 + (-10.0); };
   std::vector<double> in_lower_limits{0, 0, 0};
